% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_linear.R
\name{basic_linear}
\alias{basic_linear}
\title{basic_linear processing the raw data}
\usage{
basic_linear(Y, X, dataset, intercept = TRUE)
}
\arguments{
\item{Y}{Output variable in Equation :Y ~ X1 + X2... + Xn in character.}

\item{X}{The vector of interested variables in Equation :Y ~ X1 + X2... + Xn in character.}

\item{dataset}{Data frame containing the variables and response variable as column names.}

\item{intercept}{The indicator function of intercept. If user do not want intercept, this value is false.

#'@details basic_linear takes an equation and a data set as arguments to fit a linear regression model represented by the equation.
The data set should contain the response variable and the variables present in the equation as column names.\cr The variables are treated as continuous.\cr linear then fits the model using the method of least squares.}
}
\value{
basic_linear returns a list. The list contains the following components:\cr
#\tabular{rrrrr}{

  \emph{dataset} \tab Equation used to fit the model.\cr
  \emph{y} \tab Output variable used to fit the model.\cr
  \emph{x} \tab Interested variable used to fit the model.\cr
  \emphn{n} \tab Number of rows for x.\cr
  \emph{p} \tab Number of columns for x.\cr
  \emph{x_m} \tab The matrix of x.\cr
  \emph{y_m} \tab The matrix of y.\cr
  \emph{x_v} \tab The vector of input X separate by " ".\cr
  \emph{Y} \tab The input varaible Y.\cr
}
}
\description{
linear is used to find basic information about Simple Linear Regression and Multiple Linear Regression models with continuous variables.
}
\examples{
library(datasets)\cr
data(iris)\cr
b_l <- basic_linear("Sepal.Length", "Sepal.Width Petal.Length Species",iris)
}
