% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_Estimation.R
\name{l_Estimation}
\alias{l_Estimation}
\title{information of coefficient}
\usage{
l_Estimation(x, y, x_m, y_m, n, p)
}
\arguments{
\item{x}{Interested variable used to fit the model.}

\item{y}{Output variable used to fit the model.}

\item{x_m}{The matrix of x.}

\item{y_m}{The matrix of y.}

\item{n}{Number of rows for x.}

\item{p}{Number of columns for x.}
}
\value{
l_Estimation returns a list. The list contains the following components:\cr
#\tabular{rrrrr}{

  \emph{betahat} \tab Coefficient estimates.\cr
  \emph{y_hat} \tab Predicted Y values.\cr
  \emph{epsilonhat} \tab The residual value.\cr
  \emph{var_betahat} \tab The matrix of variance and covaraince.\cr
  \emph{se_betahat} \tab The matrix of Standard Deviation.\cr
  \emph{sigma_squared} \tab The value of MSE.\cr
  \emph{degree} \tab The degree of freedom for t test.\cr
}
}
\description{
information of coefficient
}
\details{
l_Estimation use the output of function basic_linear to give more specific information.\cr
}
\examples{
library(datasets)\cr
data(iris)\cr
b_l <- basic_linear("Sepal.Length", "Sepal.Width Petal.Length Species",iris)
l_E <- l_Estimation(b_l[[3]],b_l[[2]],b_l[[6]],b_l[[7]],b_l[[4]],b_l[[5]])
}
