% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant_level.R
\name{significant_level}
\alias{significant_level}
\title{Check significance of p-values and give symbols}
\usage{
significant_level(p_value)
}
\arguments{
\item{p_value}{a vector of p-values.}
}
\value{
\emph{significant_level} returns a vector, where each element is the p value  with a symbol denoting it's significant.\cr
\tabular{rrrrr}{
  \strong{Symbol} \tab \strong{Meaning}\cr
  \emph{***} \tab p_value<0.001\cr
  \emph{**} \tab p_value<0.01\cr
  \emph{*} \tab p_value<0.05\cr
  \emph{.} \tab p_value<0.1
}
}
\description{
Check significance of p-values and give symbols
}
\examples{
p_array <- c(4.3253e-7,0.22312,0.00222,1)
significant_level(p_array)
}
