% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_linear.R
\name{summary_linear}
\alias{summary_linear}
\title{Create A Summary Table}
\usage{
summary_linear(b_l, l_E, l_s)
}
\arguments{
\item{b_l}{basic_linear}

\item{l_E}{information of coefficient}

\item{l_s}{statistical information}
}
\value{
All the following result from the model fit:\cr
Function call for \emph{significant_level}\cr
Summary statistics of residuals\cr
Coefficient estimates and p-values\cr
The confident interval of all variables\cr
Residual standard error\cr
Multiple R-squared & Adjusted R-squared\cr
}
\description{
Create A Summary Table
}
\examples{
library(datasets)
data(iris)
b_l <- basic_linear("Sepal.Length", "Sepal.Width Petal.Length",iris)
l_E <- l_Estimation(b_l[[3]],b_l[[2]],b_l[[6]],b_l[[7]],b_l[[4]],b_l[[5]])
l_s <- l_stat(l_E[[1]],l_E[[5]],l_E[[6]],l_E[[7]],b_l[[6]],b_l[[7]],b_l[[2]])
summary_linear(b_l,l_E,l_s)
}
